+++ packages/core/src/web/app/components/StampMakerPanel/store/utils/createExpandFilter.ts
