+++ packages/core/src/web/app/components/StampMakerPanel/store/utils/createShrinkFilter.ts
